optimizando validacion isCool
