+++ server/src/main/java/com/okta/developer/CRUDAngular5SpringBoot2/CrudAngular5SpringBoot2Application.java
@@ -1,12 +1,30 @@
 package com.okta.developer.CRUDAngular5SpringBoot2;
 
+import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+import java.util.stream.Stream;
 
 @SpringBootApplication
 public class CrudAngular5SpringBoot2Application {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CrudAngular5SpringBoot2Application.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CrudAngular5SpringBoot2Application.class, args);
+    }
+
+    @Bean
+    ApplicationRunner init(CarRepository repository) {
+        return args -> {
+            Stream.of("Ferrari", "Jaguar", "Porsche", "Lamborghini", "Bugatti",
+                    "AMC Gremlin", "Triumph Stag", "Ford Pinto", "Yugo GV")
+                    .forEach(name -> {
+                        Car car = new Car();
+                        car.setName(name);
+                        repository.save(car);
+                    });
+            repository.findAll().forEach(System.out::println);
+        };
+    }
 }
