--- server/src/main/java/com/okta/developer/CRUDAngular5SpringBoot2/DemoApplication.java
@@ -1,30 +1,0 @@
-package com.okta.developer.CRUDAngular5SpringBoot2;
-
-import org.springframework.boot.ApplicationRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-import java.util.stream.Stream;
-
-@SpringBootApplication
-public class DemoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class, args);
-    }
-
-    @Bean
-    ApplicationRunner init(CarRepository repository) {
-        return args -> {
-            Stream.of("Ferrari", "Jaguar", "Porsche", "Lamborghini", "Bugatti",
-                    "AMC Gremlin", "Triumph Stag", "Ford Pinto", "Yugo GV")
-                    .forEach(name -> {
-                        Car car = new Car();
-                        car.setName(name);
-                        repository.save(car);
-                    });
-            repository.findAll().forEach(System.out::println);
-        };
-    }
-}
